refactor(host): move per-protocol pairing flows into their mixins
The May split moved the run-mode handlers into BLEMixin/ClassicMixin
but left the first-time pairing flows in HIDHost, splitting single
flows across files: _ble_restore_or_pair lived in ble.py while
_pair_ble stayed in host.py, and _query_classic_sdp called
_parse_hid_descriptor_list across the file boundary. Move _pair_ble,
_discover_ble_hid_service and _continue_ble_after_pairing into
BLEMixin, and _pair_classic, _query_classic_sdp and
_continue_classic_after_pairing into ClassicMixin. HIDHost keeps the
cross-protocol dispatch (pair_device, continue_after_pairing). Pure
move, no behavior change.
