+++ kindle_hid_passthrough/host.py
@@ -5,16 +5,8 @@ import asyncio
 from dataclasses import dataclass
 from typing import List, Optional
 
-from bumble.core import BT_BR_EDR_TRANSPORT, BT_HUMAN_INTERFACE_DEVICE_SERVICE, InvalidStateError, TimeoutError as BumbleTimeoutError
-from bumble.device import Peer
-from bumble.gatt import (
-    GATT_HUMAN_INTERFACE_DEVICE_SERVICE,
-    GATT_REPORT_CHARACTERISTIC,
-    GATT_REPORT_MAP_CHARACTERISTIC,
-)
-from bumble.hci import Address, HCI_LE_SET_PRIVACY_MODE_COMMAND, HCI_LE_Set_Privacy_Mode_Command, HCI_Write_Class_Of_Device_Command, HCI_Write_Local_Name_Command, OwnAddressType
-from bumble.hid import Host as BumbleHIDHost
-from bumble.sdp import Client as SDPClient
+from bumble.core import InvalidStateError
+from bumble.hci import HCI_LE_SET_PRIVACY_MODE_COMMAND, HCI_LE_Set_Privacy_Mode_Command, HCI_Write_Class_Of_Device_Command, HCI_Write_Local_Name_Command
 
 from ble import BLEMixin
 from classic import ClassicMixin
@@ -303,208 +295,6 @@ class HIDHost(ClassicMixin, BLEMixin):
         else:
             return await self._pair_ble(address)
 
-    async def _pair_ble(self, address: str) -> bool:
-        """Pair with a BLE device."""
-        log.info(f"[BLE] Pairing with {address}...")
-
-        target = Address(address)
-        try:
-            self.connection = await self.device.connect(
-                target,
-                own_address_type=OwnAddressType.PUBLIC,
-                timeout=config.connect_timeout,
-            )
-        except Exception as e:
-            log.error(f"[BLE] Connection failed: {e}")
-            return False
-
-        self.peer = Peer(self.connection)
-        self.current_device_address = address
-        self.connected_protocol = Protocol.BLE
-        log.success(f"[BLE] Connected to {address}")
-
-        try:
-            log.info("[BLE] Initiating pairing...")
-            await self.connection.pair()
-            log.success("[BLE] Pairing complete!")
-
-            if not self.connection.is_encrypted:
-                raise InvalidStateError("[BLE] Link not encrypted after pairing")
-
-            await self._discover_ble_hid_service()
-
-            return True
-        except Exception as e:
-            log.error(f"[BLE] Pairing failed: {e}")
-            if self.connection:
-                try:
-                    await self.connection.disconnect()
-                except Exception:
-                    pass
-                self.connection = None
-                self.peer = None
-            return False
-
-    async def _discover_ble_hid_service(self, process_reports: bool = False):
-        """Discover BLE GATT HID service and cache descriptor."""
-        await self.peer.discover_services()
-
-        if not self.device_name:
-            await self._read_ble_device_name()
-
-        hid_services = [s for s in self.peer.services if s.uuid == GATT_HUMAN_INTERFACE_DEVICE_SERVICE]
-        if not hid_services:
-            if process_reports:
-                raise InvalidStateError("[BLE] HID service not found")
-            log.warning("[BLE] HID service not found")
-            return
-
-        hid_service = hid_services[0]
-        log.success("[BLE] Found HID service")
-
-        await self.peer.discover_characteristics(service=hid_service)
-
-        for char in hid_service.characteristics:
-            if char.uuid == GATT_REPORT_MAP_CHARACTERISTIC and not self.report_map:
-                try:
-                    value = await self.peer.read_value(char)
-                    self.report_map = bytes(value)
-                    log.success(f"[BLE] Got descriptor: {len(self.report_map)} bytes")
-
-                    address = self.current_device_address
-                    self.device_cache.save(address, {
-                        'report_map': self.report_map.hex(),
-                        'device_name': self.device_name
-                    })
-                except Exception as e:
-                    log.warning(f"[BLE] Failed to read report map: {e}")
-
-            elif process_reports and char.uuid == GATT_REPORT_CHARACTERISTIC:
-                await self._process_ble_report_char(char)
-
-    async def _pair_classic(self, address: str) -> bool:
-        """Pair with a Classic Bluetooth device."""
-        log.info(f"[Classic] Pairing with {address}...")
-
-        try:
-            target_address = Address(address, Address.PUBLIC_DEVICE_ADDRESS)
-            self.connection = await self.device.connect(
-                target_address,
-                transport=BT_BR_EDR_TRANSPORT,
-                timeout=config.connect_timeout,
-            )
-            log.success(f"[Classic] Connected to {address}")
-        except (asyncio.TimeoutError, BumbleTimeoutError):
-            log.error(f"[Classic] Connection timeout after {config.connect_timeout}s")
-            return False
-        except Exception as e:
-            log.error(f"[Classic] Connection failed: {e}")
-            return False
-
-        self.current_device_address = address
-        self.connected_protocol = Protocol.CLASSIC
-
-        link_key_received = asyncio.Event()
-
-        def on_device_link_key(_bd_addr, link_key, key_type):
-            log.success(f"[Classic] Link key received: type={key_type}")
-            link_key_received.set()
-
-        self.device.host.on('link_key', on_device_link_key)
-
-        try:
-            log.info("[Classic] Authenticating...")
-            try:
-                await asyncio.wait_for(self.connection.authenticate(), timeout=30.0)
-                log.success("[Classic] Authentication complete")
-            except Exception as e:
-                log.warning(f"[Classic] Authentication: {e}")
-
-            log.info("[Classic] Waiting for link key...")
-            try:
-                await asyncio.wait_for(link_key_received.wait(), timeout=5.0)
-                log.success("[Classic] Link key saved")
-            except asyncio.TimeoutError:
-                log.warning("[Classic] Link key event timeout (may already be saved)")
-
-            if not self.connection.is_encrypted:
-                log.info("[Classic] Requesting encryption...")
-                try:
-                    await asyncio.wait_for(
-                        self.connection.encrypt(enable=True),
-                        timeout=10.0
-                    )
-                except Exception as e:
-                    log.warning(f"[Classic] Encryption: {e}")
-
-            await self._query_classic_sdp(address)
-
-            if self.keystore:
-                keys = await self.keystore.get(address)
-                if keys and keys.link_key:
-                    log.success("[Classic] Link key verified")
-                else:
-                    log.warning("[Classic] Link key not found in keystore!")
-
-            self.device.host.remove_listener('link_key', on_device_link_key)
-            return True
-
-        except Exception as e:
-            log.error(f"[Classic] Pairing failed: {e}")
-            self.device.host.remove_listener('link_key', on_device_link_key)
-            if self.connection:
-                try:
-                    await self.connection.disconnect()
-                except Exception:
-                    pass
-                self.connection = None
-            return False
-
-    async def _query_classic_sdp(self, address: str = None):
-        """Query SDP for HID descriptor and cache it."""
-        if not self.connection:
-            return
-
-        address = address or self.current_device_address
-
-        log.info("[Classic] Querying SDP...")
-        try:
-            sdp_client = SDPClient(self.connection)
-            await asyncio.wait_for(sdp_client.connect(), timeout=5.0)
-
-            result = await asyncio.wait_for(
-                sdp_client.search_attributes(
-                    [BT_HUMAN_INTERFACE_DEVICE_SERVICE],
-                    [0x0100, 0x0206]
-                ),
-                timeout=10.0
-            )
-
-            if result:
-                for record in result:
-                    for attr in record:
-                        if hasattr(attr, 'id') and attr.id == 0x0206:
-                            self._parse_hid_descriptor_list(attr.value)
-                        elif hasattr(attr, 'id') and attr.id == 0x0100:
-                            try:
-                                name = attr.value.value
-                                if isinstance(name, bytes):
-                                    name = name.decode('utf-8', errors='replace')
-                                self.device_name = str(name)
-                            except Exception:
-                                pass
-
-            await sdp_client.disconnect()
-
-            if self.report_map:
-                self.device_cache.save(address, {
-                    'report_map': self.report_map.hex(),
-                    'device_name': self.device_name or 'Unknown'
-                })
-                log.success(f"[Classic] Cached descriptor ({len(self.report_map)} bytes)")
-        except Exception as e:
-            log.warning(f"[Classic] SDP query failed: {e}")
-
     async def continue_after_pairing(self):
         """Continue into run mode after successful pairing."""
         if not self.connected_protocol:
@@ -527,57 +317,6 @@ class HIDHost(ClassicMixin, BLEMixin):
         log.success(f"\n[{proto_name}] Paired and receiving HID reports. Press Ctrl+C to exit.")
 
         await self._disconnection_event.wait()
-
-    async def _continue_classic_after_pairing(self):
-        """Continue Classic connection after pairing."""
-        self.hid_host = BumbleHIDHost(self.device)
-        self.hid_host.on(BumbleHIDHost.EVENT_INTERRUPT_DATA, self._on_classic_interrupt_data)
-        self.hid_host.on(BumbleHIDHost.EVENT_VIRTUAL_CABLE_UNPLUG, self._on_virtual_cable_unplug)
-        log.info("[Classic] HID Host created")
-
-        self.hid_host.on_device_connection(self.connection)
-
-        log.info("[Classic] Connecting to HID control channel...")
-        try:
-            await asyncio.wait_for(self.hid_host.connect_control_channel(), timeout=5.0)
-            log.success("[Classic] HID control channel connected")
-        except Exception as e:
-            log.warning(f"[Classic] Control channel: {e}")
-
-        log.info("[Classic] Connecting to HID interrupt channel...")
-        try:
-            await asyncio.wait_for(self.hid_host.connect_interrupt_channel(), timeout=5.0)
-            log.success("[Classic] HID interrupt channel connected")
-        except Exception as e:
-            log.warning(f"[Classic] Interrupt channel: {e}")
-
-        if not self.hid_host.l2cap_intr_channel:
-            log.error("[Classic] Failed to connect HID interrupt channel")
-            return
-
-        self._classic_set_report_protocol()
-        self._finalize_classic_hid()
-
-    async def _continue_ble_after_pairing(self):
-        """Continue BLE connection after pairing."""
-        if not self.connection:
-            log.info(f"[BLE] Reconnecting to {self.current_device_address}...")
-            target = Address(self.current_device_address)
-            self.connection = await self.device.connect(
-                target,
-                own_address_type=OwnAddressType.PUBLIC,
-                timeout=config.connect_timeout,
-            )
-            self.peer = Peer(self.connection)
-            self.connection.on('disconnection', self._on_disconnection)
-            await self._ble_restore_or_pair()
-        else:
-            log.info("[BLE] Using existing connection from pairing")
-            if not self.peer:
-                self.peer = Peer(self.connection)
-            await self._ble_restore_or_pair()
-
-        await self._setup_ble_hid()
 
     # ==================== COMMON ====================
